pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-vault"
version = "0.1.0"
description = "A tool for managing Agent Skills across different AI frameworks"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Pascal"}
]
keywords = ["agent", "skills", "ai", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pathspec>=0.11.0",
    "inquirerpy>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
skill-vault = "skill_vault.cli:main"
sv = "skill_vault.cli:main"
skill-eye = "skill_vault.cli:main"

[project.urls]
Homepage = "https://github.com/pascal/skill-vault"
Repository = "https://github.com/pascal/skill-vault"
Issues = "https://github.com/pascal/skill-vault/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
skill_vault = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
